Añadiendo cambios a la api

+++ api_restSD/src/main/java/com/fiuni/sd/service/cliente/ClienteServiceImpl.java
@@ -1,11 +1,6 @@
 package com.fiuni.sd.service.cliente;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
@@ -14,51 +9,54 @@ import com.fiuni.sd.domain.cliente.ClienteDomain;
 import com.fiuni.sd.dto.cliente.ClienteDTO;
 import com.fiuni.sd.dto.cliente.ClienteResult;
 import com.fiuni.sd.service.base.BaseServiceImpl;
+import com.fiuni.sd.utils.ResourceNotFoundException;
 
 @Service
-public class ClienteServiceImpl  extends BaseServiceImpl<ClienteDTO, ClienteDomain, ClienteResult> implements IClienteService{
-	
+public class ClienteServiceImpl extends BaseServiceImpl<ClienteDTO, ClienteDomain, ClienteResult>
+		implements IClienteService {
+
 	@Autowired
-	private IClienteDao clientDao;
+	private IClienteDao clientDao; // repository
 
 	@Override
-	public ClienteDTO save(ClienteDTO dto) {
-		final ClienteDomain clientDomain = convertDtoToDomain(dto);
-		final ClienteDomain client = clientDao.save(clientDomain);
-		return convertDomainToDto(client);
+	public ClienteDTO save(final ClienteDTO dto) {
+		return convertDomainToDto(clientDao.save(convertDtoToDomain(dto)));
 	}
 
 	@Override
-	public ClienteDTO getById(Integer id) {
-		Optional<ClienteDomain> result = clientDao.findById(id);
-        ClienteDTO client = null;
-        if(result.isPresent()){
-        	client = convertDomainToDto(result.get());
-        } else {
-            throw new RuntimeException("Did not find client id: " + id);
-        }
-        return client;
+	public ClienteDTO getById(final Integer id) {
+		return clientDao.findById(id)//
+				.map(this::convertDomainToDto)//
+				.orElseThrow(() -> new ResourceNotFoundException("Cliente", "id", id));
 	}
 
 	@Override
-	public ClienteResult getAll(Pageable pageable) {
-		final List<ClienteDTO> clients = new ArrayList<>();
-		Page<ClienteDomain> results = clientDao.findAll(pageable);
-		for (ClienteDomain clientDomain : results) {
-			clients.add(convertDomainToDto(clientDomain));
+	public ClienteResult getAll(final Pageable pageable) {
+		final ClienteResult result = new ClienteResult();
+		result.setClientes(clientDao.findAll(pageable)//
+				.map(this::convertDomainToDto)//
+				.toList());
+		return result;
+	}
+
+	@Override
+	public void deleteById(final Integer id) {
+		if (!clientDao.existsById(id)) {
+			throw new ResourceNotFoundException("Cliente", "id", id);
 		}
-		ClienteResult clientResult = new ClienteResult();
-		clientResult.setClientes(clients);
-		return clientResult;
-	}
-
-	@Override
-	public void deleteById(int id) {
 		clientDao.deleteById(id);
 	}
 
 	@Override
-	protected ClienteDTO convertDomainToDto(ClienteDomain domain) {
+	public ClienteDTO update(final Integer id, final ClienteDTO dto) {
+		if (!clientDao.existsById(id)) {
+			throw new ResourceNotFoundException("Cliente", "id", id);
+		}
+		return convertDomainToDto(clientDao.save(convertDtoToDomain(dto)));
+	}
+
+	@Override
+	protected ClienteDTO convertDomainToDto(final ClienteDomain domain) {
 		final ClienteDTO dto = new ClienteDTO();
 		dto.setId(domain.getId());
 		dto.setCi(domain.getCi());
@@ -66,11 +64,11 @@ public class ClienteServiceImpl  extends BaseServiceImpl<ClienteDTO, ClienteDoma
 		dto.setDireccion(domain.getDireccion());
 		dto.setEmail(domain.getEmail());
 		dto.setTelefono(domain.getTelefono());
-        return dto;
+		return dto;
 	}
 
 	@Override
-	protected ClienteDomain convertDtoToDomain(ClienteDTO dto) {
+	protected ClienteDomain convertDtoToDomain(final ClienteDTO dto) {
 		final ClienteDomain domain = new ClienteDomain();
 		domain.setId(dto.getId());
 		domain.setCi(dto.getCi());
@@ -78,7 +76,7 @@ public class ClienteServiceImpl  extends BaseServiceImpl<ClienteDTO, ClienteDoma
 		domain.setDireccion(dto.getDireccion());
 		domain.setEmail(dto.getEmail());
 		domain.setTelefono(dto.getTelefono());
-        return domain;
+		return domain;
 	}
 
 }
